pyproject.toml
[project]
name = "fastapi-logging-samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "prometheus-client>=0.23.1",
]
